refactor: call PLENA compiler directly from ATen frontend
