+++ aten/ops/plena/attention_ops.py
@@ -1,201 +1,5 @@
-"""PLENA backend implementation for Flash Attention operator."""
+"""PLENA backend compatibility shim for Flash Attention."""
 
 
 def flash_attention_plena(prog, Q, K, V, scale=None, hq=1, hkv=1, h_qkv=None, causal_mask=None):
-    """PLENA backend: Flash Attention via PlenaCompiler.
-
-    Dispatches to one of two codegen paths based on shape:
-      * MHA  (hq == hkv == 1) — online-softmax loop using PlenaCompiler primitives
-      * GQA  (hq // hkv > 1)  — fused codegen via `flash_attn_asm` template, packs
-        `hq/hkv` Q heads into blen of M_BTMM (hardware GQA fusion, matches main)
-
-    Args:
-        prog:   PlenaCompiler instance
-        Q:      VRAMMatrixVar — Q in VRAM, shape (seq_len, hq*h_qkv)
-        K:      InputVar — K in HBM, shape (seq_len, hkv*h_qkv_padded)
-        V:      InputVar — V in HBM, shape (seq_len, hkv*h_qkv_padded)
-        scale:  Attention scale (default 1/sqrt(head_dim))
-        hq, hkv, h_qkv: GQA params. Defaults treat input as single-head MHA.
-        causal_mask: VRAMMatrixVar or None — (mlen, mlen) mask with 0 on/below
-                     diagonal and -inf above. Added to S before softmax.
-
-    Returns:
-        VRAMMatrixVar for O, shape matching Q.
-    """
-
-    # Detect MHA vs GQA
-    if hq == 1 and hkv == 1:
-        return _flash_attention_mha(prog, Q, K, V, scale, causal_mask=causal_mask)
-
-    # GQA: dispatch to fused codegen
-    if h_qkv is None:
-        raise ValueError("GQA mode requires h_qkv to be specified")
-    if causal_mask is not None:
-        raise NotImplementedError("causal_mask is not yet supported for GQA flash attention")
-    return _flash_attention_gqa_fused(prog, Q, K, V, scale, hq, hkv, h_qkv)
-
-
-def _flash_attention_mha(prog, Q, K, V, scale, causal_mask=None):
-    """Single-head flash attention using PlenaCompiler primitives (online softmax).
-
-    Args:
-        prog:         PlenaCompiler instance
-        Q:            VRAMMatrixVar — query matrix in VRAM
-        K:            InputVar — key matrix in HBM
-        V:            InputVar — value matrix in HBM
-        scale:        Attention scale factor
-        causal_mask:  VRAMMatrixVar or None — (mlen, mlen) mask with 0 on/below
-                      diagonal and -inf above. When provided, added to S_block
-                      after QK^T and before online softmax.
-    """
-    import math
-
-    seq_len, head_dim = Q.shape
-    mlen = prog.mlen
-
-    if scale is None:
-        scale = 1.0 / math.sqrt(head_dim)
-
-    num_q_blocks = seq_len // mlen
-    num_k_blocks = seq_len // mlen
-
-    S_block = prog.alloc("S", mlen, mlen)
-    PV = prog.alloc("PV", mlen, head_dim)
-    O = prog.alloc("O", seq_len, head_dim)
-
-    for q_idx in range(num_q_blocks):
-        prog.init_online_softmax(q_idx, O)
-
-        for k_idx in range(num_k_blocks):
-            prog.vram_sub_projection_T_to(
-                Q,
-                q_idx,
-                K,
-                k_idx,
-                S_block,
-                target_row_idx=0,
-                target_col_idx=0,
-            )
-            # Apply causal mask before softmax: S += mask (-inf above diagonal)
-            if causal_mask is not None:
-                prog.vram_add(S_block, causal_mask)
-            prog.online_softmax_block(S_block, scale)
-            prog.compute_pv(S_block, V, k_idx, PV, head_dim)
-            prog.scale_o_row(O, q_idx)
-            prog.vram_add(O, PV, dst_row_offset=q_idx * mlen)
-
-        prog.final_scale_o(q_idx, O)
-
-    return O
-
-
-def _flash_attention_gqa_fused(prog, Q, K, V, scale, hq, hkv, h_qkv):
-    """GQA flash attention using fused M_BTMM codegen (main branch template).
-
-    Packs `hq/hkv` Q heads into the `blen` systolic dimension — one M_BTMM
-    produces `ratio` head outputs in parallel.  Requires:
-      (hq // hkv) == prog.blen, and (hq // hkv) * h_qkv == prog.mlen.
-    """
-    import math
-    from compiler.asm_templates.flashattn import flash_attn_asm
-    from compiler.asm_templates import preload_addr_reg_asm
-
-    ratio = hq // hkv
-    mlen = prog.mlen
-    blen = prog.blen
-    vlen = mlen
-
-    if ratio != blen:
-        raise ValueError(
-            f"GQA ratio hq/hkv={ratio} must equal blen={blen} (hardware packs "
-            f"heads into blen).  Use ratio=4 with default blen."
-        )
-    if ratio * h_qkv != mlen:
-        raise ValueError(
-            f"GQA constraint: (hq/hkv)*h_qkv = {ratio * h_qkv} must equal "
-            f"mlen={mlen}.  E.g. hq=4, hkv=1, h_qkv=16 with mlen=64."
-        )
-
-    s_q, _q_total_dim = Q.shape
-    s_kv, _k_total_dim = K.shape
-
-    if scale is None:
-        scale = 1.0 / math.sqrt(h_qkv)
-
-    # Allocate HBM addr registers for K and V (C_SET_ADDR_REG aN)
-    # Make sure K/V HBM subst-matrix registry knows about them
-    prog._ensure_hbm_sub_matrix_registered(K)
-    prog._ensure_hbm_sub_matrix_registered(V)
-    alloc = prog.register_allocator
-    k_addr, v_addr = alloc.allocate_addr(2)
-    gp_for_preload = alloc.allocate_gp(2)
-    setup = preload_addr_reg_asm(
-        addr_reg_to_set=[k_addr, v_addr],
-        available_registers=gp_for_preload,
-        addr_reg_val=[K.hbm_addr, V.hbm_addr],
-    )
-    alloc.free_gp(gp_for_preload)
-    prog.emit(setup)
-
-    # Allocate VRAM buffers mirroring main's layout.
-    #   S, PV each require mlen*mlen*ratio elements; O is s_q * (hq*h_qkv).
-    # We let PlenaCompiler's allocator handle placement — main's template uses
-    # vector_sram_base_address + computed offsets for S/PV/O, so they must be
-    # allocated contiguously starting right after Q.  allocate_vram_matrix
-    # bump-allocates, which preserves that contiguity.
-    q_vram_base = prog.get_vram_addr(Q.name)
-    s_name = prog._scoped_name("_gqa_S")
-    pv_name = prog._scoped_name("_gqa_PV")
-    o_name = prog._scoped_name("O")
-
-    # Express sizes as (rows, cols) that multiply to the required counts.
-    prog.allocate_vram_matrix(name=s_name, rows=mlen * ratio, cols=mlen, strict=False)
-    prog.allocate_vram_matrix(name=pv_name, rows=mlen * ratio, cols=mlen, strict=False)
-    prog.allocate_vram_matrix(name=o_name, rows=s_q, cols=hq * h_qkv, strict=False)
-
-    # Reserve FPRAM for multi-head softmax state.  main's flash_attn_asm
-    # assumes slots 0..2 hold constants (0.0, scale, -inf, preloaded by the
-    # test harness) and softmax state starts at fp_sram_start_address, with
-    # 3 triples (m_old, m_res, l_old) per head, strided 3*br apart.
-    # Reserve slots 0..2 first so our bump-allocated state lives at offset 3+.
-    br = min(mlen, s_q)
-    fp_allocs = prog.fpram_allocator
-    if "_gqa_fp_const_zero" not in fp_allocs.allocations:
-        fp_allocs.allocate(name="_gqa_fp_const_zero", size=1)
-        fp_allocs.allocate(name="_gqa_fp_const_scale", size=1)
-        fp_allocs.allocate(name="_gqa_fp_const_neg_inf", size=1)
-    fp_state_size = 3 * br * ratio
-    fp_info = prog.add_fpram_object(name="_gqa_softmax_state", size=fp_state_size)
-    if fp_info.fpram_addr is None:
-        raise RuntimeError("Failed to allocate FPRAM for GQA softmax state")
-    fp_start = fp_info.fpram_addr
-
-    # Call main's fused GQA template
-    asm = flash_attn_asm(
-        mlen=mlen,
-        vlen=vlen,
-        blen=blen,
-        batch=1,
-        hq=hq,
-        hkv=hkv,
-        d=h_qkv,
-        q_len=s_q,
-        kv_len=s_kv,
-        alive_registers_int=list(range(1, 16)),
-        alive_registers_fp=list(range(1, 8)),
-        vector_sram_base_address=q_vram_base,
-        fp_sram_start_address=fp_start,
-        k_base_hbm_offset_reg=k_addr,
-        v_base_hbm_offset_reg=v_addr,
-    )
-    prog.emit(asm)
-
-    # Release HBM addr regs (they're only needed during the call)
-    alloc.free_addr([k_addr, v_addr])
-
-    # Return O as a VRAMMatrixVar the caller can consume
-    from compiler.aten.plena.vars import VRAMMatrixVar
-
-    O = VRAMMatrixVar(prog, o_name, (s_q, hq * h_qkv), display_name="O")
-    prog._tensors[o_name] = O
-    return O
+    return prog.flash_attention(Q, K, V, scale=scale, hq=hq, hkv=hkv, h_qkv=h_qkv, causal_mask=causal_mask)
